Move classes for Unnest Operator to a separate package
