+++ presto-main/src/main/java/io/prestosql/operator/unnest/ArrayOfRowsUnnester.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.prestosql.operator;
+package io.prestosql.operator.unnest;
 
 import com.google.common.collect.ImmutableList;
 import io.prestosql.spi.PageBuilder;
