+++ presto-main/src/main/java/io/prestosql/operator/unnest/MapUnnester.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.prestosql.operator;
+package io.prestosql.operator.unnest;
 
 import io.prestosql.spi.PageBuilder;
 import io.prestosql.spi.block.Block;
@@ -23,18 +23,20 @@ import javax.annotation.Nullable;
 import static com.google.common.base.Preconditions.checkState;
 import static java.util.Objects.requireNonNull;
 
-public class ArrayUnnester
+public class MapUnnester
         implements Unnester
 {
-    private final Type elementType;
-    private Block arrayBlock;
+    private final Type keyType;
+    private final Type valueType;
+    private Block block;
 
     private int position;
     private int positionCount;
 
-    public ArrayUnnester(Type elementType)
+    public MapUnnester(Type keyType, Type valueType)
     {
-        this.elementType = requireNonNull(elementType, "elementType is null");
+        this.keyType = requireNonNull(keyType, "keyType is null");
+        this.valueType = requireNonNull(valueType, "valueType is null");
     }
 
     @Override
@@ -46,23 +48,24 @@ public class ArrayUnnester
     @Override
     public final int getChannelCount()
     {
-        return 1;
+        return 2;
     }
 
     @Override
     public final void appendNext(PageBuilder pageBuilder, int outputChannelOffset)
     {
-        checkState(arrayBlock != null, "arrayBlock is null");
-        BlockBuilder blockBuilder = pageBuilder.getBlockBuilder(outputChannelOffset);
-        elementType.appendTo(arrayBlock, position, blockBuilder);
-        position++;
+        checkState(block != null, "block is null");
+        BlockBuilder keyBlockBuilder = pageBuilder.getBlockBuilder(outputChannelOffset);
+        BlockBuilder valueBlockBuilder = pageBuilder.getBlockBuilder(outputChannelOffset + 1);
+        keyType.appendTo(block, position++, keyBlockBuilder);
+        valueType.appendTo(block, position++, valueBlockBuilder);
     }
 
     @Override
-    public void setBlock(@Nullable Block arrayBlock)
+    public void setBlock(@Nullable Block mapBlock)
     {
-        this.arrayBlock = arrayBlock;
+        this.block = mapBlock;
         this.position = 0;
-        this.positionCount = arrayBlock == null ? 0 : arrayBlock.getPositionCount();
+        this.positionCount = mapBlock == null ? 0 : mapBlock.getPositionCount();
     }
 }
