+++ presto-main/src/main/java/io/prestosql/operator/unnest/Unnester.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.prestosql.operator;
+package io.prestosql.operator.unnest;
 
 import io.prestosql.spi.PageBuilder;
 import io.prestosql.spi.block.Block;
