+++ presto-main/src/main/java/io/prestosql/sql/planner/LocalExecutionPlanner.java
@@ -243,8 +243,8 @@ import static io.prestosql.operator.TableWriterOperator.FRAGMENT_CHANNEL;
 import static io.prestosql.operator.TableWriterOperator.ROW_COUNT_CHANNEL;
 import static io.prestosql.operator.TableWriterOperator.STATS_START_CHANNEL;
 import static io.prestosql.operator.TableWriterOperator.TableWriterOperatorFactory;
-import static io.prestosql.operator.UnnestOperator.UnnestOperatorFactory;
 import static io.prestosql.operator.WindowFunctionDefinition.window;
+import static io.prestosql.operator.unnest.UnnestOperator.UnnestOperatorFactory;
 import static io.prestosql.spi.StandardErrorCode.COMPILER_ERROR;
 import static io.prestosql.spi.type.BigintType.BIGINT;
 import static io.prestosql.spi.type.TypeUtils.writeNativeValue;
