+++ presto-main/src/test/java/io/prestosql/operator/unnest/TestUnnestOperator.java
@@ -11,11 +11,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.prestosql.operator;
+package io.prestosql.operator.unnest;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import io.prestosql.metadata.Metadata;
+import io.prestosql.operator.DriverContext;
+import io.prestosql.operator.OperatorFactory;
 import io.prestosql.spi.Page;
 import io.prestosql.spi.type.ArrayType;
 import io.prestosql.spi.type.RowType;
